New target board NUCLEO144-F746ZG for ChibiOS (V2)
Support for the F7 board in the overlay concept was added, here are the
additional (target) files for the above mentioned board. This is initial
work so there will be corrections to make.

- initiated a new target board and added files accordingly
- additional files made to make use of the overlay concept
- changed cmake files to match board specifics
- added all copyright headers
- added empty end lines to make Git diff happy

Target has been build, compiled and deployed successfully to the board.
A larson scanner should be shown once deployed.

Signed-off-by: Peter Wessel piwi1263@gmail.com