Windows.Devices.Gpio POC
This is work in progress for evaluation only.
Goal here is to check if we are going the right way of doing things regarding file locations, methods and so on. Code is really incomplete and many methods do not contain any safety checks.

This PR also needs changes in nf-class-libraries projects.

Signed-off-by: Christophe Gerbier <christophe@mikrobusnet.org>