Work on new DateTime native side
Removal of TimeZone (see issue #335)
Rework on DateTime structure to remove method related to UTC.local conversions, since nanoFramework is only using UTC.
Moved some managed methods to the native side, using ChibiOS API where possible.

This PR is meant to be used with the changes made in nf-class-libraries regarding DateTime as well. The PR in nf-class-libraries is the managed counterpart of this one.

Signed-off-by: Christophe Gerbier <christophe@mikrobusnet.org>