Work on Execution
The CLR Interpreter runs on it's own OS thread (then jumping from one nanoFramework thread to the next) it leaves no time slot for other OS threads to execute on OSes that relly on cooperative execution (menaning that a thread has to relinquish execution control to the OS so other threads can run). Because of this, on nF thread switching, there is a call to NANOCLR_RELINQUISHEXECUTIONCONTROL to allow this to happen on OSes that require that. ChibiOS is one of those when it's running on cooperative mode.

Signed-off-by: José Simões <jose.simoes@eclo.solutions>