Inital Submit of ESP32 support
This builds the debug build for NanoClr including the following class drivers:-

Gpio/ Events ( Working )
Spi ( In the process of being tested )
I2c ( Working with OLED display )

There is an initial document on building the NanoClr in the targets\freertos\Esp32\BUILD-esp32.md

The build creates a new partition table with a special type '84" partition for the deployment area.
The build links with pre built IDF libraries that can be built from the IDF sample "Blinky".

The NanoCLr will not work with the current Visual.Studio.extension until a small patch is implemented to allow all serial ports to be seen and to set serial port to 8 bit data ( it defaults to 7 bit with the ESP32 DevKitc board)

Currently no networking although the Networking is inialized at interally at the startup. There is some code to initlaise the Smart config but this has been disabled as it was causing an exception. There is also an incomplete implementation of the Windows.Devices.Wifi.

Currently haven't seen any need for the NanoBooter, although there is commented out statements in the CmakeLists.txt to build an incomplete version of it.

Still a lot to be done.

* Networking Tcp/ip, Udp, Bluetooth
* Other device drivers pwm, adc etc
* Power management
* EPS32 pin mapping
* Other EPS32 / IDF specific functions
* Managed code libaries for ESP32 specific functions

Other issues
FPU is currently emulated. Doesn't use RTC yet.
The Release build currently doesn't build.