Include community board in select
The compile breaks because of undeclared vars in I2C version for F4-Series boards. For this series version 1 has to be forced to use at two places in the cmake module.

By adding the community board string to the distinct list of F4 boards that select the correct version for I2C the compiler was happy and builds successfully.

Signed-off-by: piwi1263 <piwi1263@gmail.com>