Update win_dev_gpio_native_Windows_Devices_Gpio_GpioPin.cpp
"state" variable was not used.
No need to get the "ioLine" variable if we don't need it. So moved it to the location where the code executes. 

Those two changes cost less CPU time and memory. A few more bytes could probably be saved by not declaring the "ioLine" variable.