Added SUPPORT_ANY_BASE_CONVERSION
I found that we can set the option SUPPORT_ANY_BASE_CONVERSION for the ESP32. .NET Framework supports base 2, 8, 10 and 16 conversions. And the ESP32 functions strtoll and strtoull which are used in NativeToInt64 are also supporting base 2, 8, 10 and 16 conversions. Additionally the strtod function that is used in NativeToDouble is also working. The binary sizes of nanoCLR.bin for MinSizeRel builds with or without this option are the same.
By setting these option the ESP32 supports also the base 2 and 8 conversion.