Removed setting default time and date during MCU startup.
I removed section with setting default time and date. 
This is initial implementation. For proper work of RTC additional configuration of SNVS_LP is necessary and it will be added later.
With SNVS_LP time will be measured with low power section of the chip and it should work properly even after MCU reset, so saving and restoring might be not necessary.