Use HSE for RTC
Changed RTC clock source to use HSE since LSE is not available and LSI is to inaccurate.
Two new defines are needed for clock prescaler. See AN4759 on page 9 and 10 for more details.
It has to be noted that ChibiOS does auto reduce the divisors for async and sync prescaler.
We must use 124+1 and 7999+1 for the prescaler defines as can be seen in the changed file.

Signed-off-by: piwi1263 <piwi1263@gmail.com>