Update cpu_gpio.cpp
As part of solving issue #705 it was noticed that, if any INPUT pins are defined, the application raises an unhandled exception when entering deep sleep state.
This change is to fix this problem.
It was found that the way <<CPU_GPIO_Uninitialize>> is implemented, a linked list itteration fails after the current item in the linked list gets removed during the itteration. 
A better coding pattern is now used which allows removal of the item while itterating.